Nicer implementation
Added WinIcon file (actually changed the name from MyIcon to WinIcon,
but I also changed some parts of the code)

Removed Clickable.js -> this is a functional language, why would I
encapsulate? Actually I made a mistake doing that, when creating a
Clickable object, then creating an Icon on top of that, the functions
mouse_down() weren't aware of the existence of the Icon, which is kind
of the whole point.

Now Windows open files instead of raw HTML strings, which makes
everything nicer.

Added some comments.