Small changes and behavioral bug fixes
Added a title for a WinIcon and a smoother hover background, with
roundish edges.
Fixed the bug where when you extend a page, it doesn't extend it's
content.
Fixed the "focus" problem because it was setting cur = $(this.iDiv)
inside focus, and it had nothing to do here.
Created another test page which will be nicely formatted later.